pyproject.toml
[tool.poetry]
name = "creating-api"
version = "0.1.0"
description = ""
authors = ["Caio Machado <moliveiracaio@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]
    
[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
